improved_project/folder_api/requirements.txt
fastapi
uvicorn
aio_pika
redis
python-dotenv
pydantic
